+++ src/views/link.js
@@ -3,23 +3,33 @@ import usage from './link.usage.js';
 
 export default function(discovery) {
     discovery.view.define('link', function(el, config, data, context) {
-        const { content } = config;
+        const { content, onClick } = config;
         let { href, text, external } = data || {};
 
         if (typeof data === 'string') {
             href = text = data;
         }
 
-        if (!text && href) {
+        if (text === undefined && href) {
             text = href;
-        } else if (!href && text) {
+        } else if (href === undefined && text) {
             href = text;
         }
 
-        el.href = href;
+        if (href) {
+            el.href = href;
+        }
 
         if (external) {
             el.setAttribute('target', '_blank');
+        }
+
+        if (typeof onClick === 'function') {
+            el.classList.add('onclick');
+            el.addEventListener('click', (e) => {
+                e.preventDefault();
+                onClick(el, data, context);
+            });
         }
 
         if (content) {